.github: notify on merge conflicts in PRs
Add a pipeline that's checking for and notifying on merge conflicts
within open pull requests. It's triggered whenever one of the baseline
branches (eg. master, maint-*, etc) are pushed.

Unfortunately, we don't have any way for rate limiting within a single
PR yet, so on a day with lots of merges, it could become a bit spammy.

Signed-off-by: Enrico Weigelt, metux IT consult <info@metux.net>