hw/modesetting: Fix up cursor size probing.
modesetting has a sort of broken concept of getting the cursor dimensions since 1f41320,
which causes issues on virtual machines and older AMD hardware.

Our solution is to use SIZE_HINTS which is available in i915 since 2023,
otherwise fallback to getting the cursor size provided by the driver.

If a driver wants to have a more optimal cursor size,
it is more than welcome to implement SIZE_HINTS in it's DRM interface.

Taken from my own fork of Xorg.

v2: Fix some bugs in the implemenation.