modesetting: Only paint as much of the cursor image as needed
We often use very large cursor sizes, compared to the sizes of the
curor glyphs, e.g. 64x64 cursors and 16x16 or smaller glyphs.

This patch makes is so that we only paint the size of the new
cursor glyph and clear the old one if needed, instead of painting
the entire cursor buffer, which is often far larger than the glyphs.

Signed-off-by: stefan11111 <stefan11111@shitposting.expert>