modesetting: rename OPTION_FALLBACK_CURSOR to OPTION_CURSOR_SIZE
Now that we probe for the cursor size automatically, even if
SIZE_HINTS isn't available, it isn't very useful to set the
size of the fallback cursor.

However, it might still be useful to force a particular cursor
size via xorg.conf.
For example, if on a system the probes fail, or if the user wants
a particular cursor size, that is higher that the minimum size.

Signed-off-by: stefan11111 <stefan11111@shitposting.expert>