hw/xfree86/common: revert: https://github.com/X11Libre/xserver/commit/d09b3dae3e91bcc5c465c827950bcd40c1988093
Drivers like xf86-video-fbdev can drive almost anything.
This makes it an excellent fallback driver, useful when the
proper driver breaks, or on livecd's.

Sadly, this also means that it can collide with other drivers,
using the same pci slot as them.
This usually doesn't lead to broken graphics, since the dedicated
drivers usually take priority, and the device ignores the fbdev interface.'

This means that when autoconfiguring, if the fbdev driver is installed,
the X server will likely use it, and that can lead to 2 X11 screens being
initializes on the same physical device, one being the dedicated driver
the user intended to use, another being fbdev.

This configuration is suboptimal and can break glx.

Also converted all of these globals to ints.

Signed-off-by: stefan11111 <stefan11111@shitposting.expert>